rename currency_formatter.rs to formatter.rs
